Avoid ROS imports for simulation script
