+++ lsy_drone_racing/utils/__init__.py
@@ -1,0 +1,13 @@
+"""Utility module.
+
+We separate utility functions that require ROS into a separate module to avoid ROS as a
+dependency for sim-only scripts.
+"""
+from lsy_drone_racing.utils.utils import (
+    check_gate_pass,
+    draw_trajectory,
+    load_config,
+    load_controller,
+)
+
+__all__ = ["load_config", "load_controller", "check_gate_pass", "draw_trajectory"]
